// src/types.ts
    export type NodeType = 'individual' | 'regional_node' | 'global_node';
    export type NodeTier = 'inference' | 'aggregator' | 'training' | 'feedback';